+++ src/constants.js
@@ -1,3 +1,12 @@
 // module name from module.json
 export const MODULE_NAME = "dungeon-draw";
+
+// set as wall flag
+export const DUNGEON_VERSION = "1.0";
+
+export const FLAG_DUNGEON_VERSION = "dungeonVersion";
+
+export const SETTING_3DCANVAS_ENABLED = "3DCanvasEnabled";
+export const SETTING_CUSTOM_THEMES = "customThemes";
+export const SETTING_RELEASE_NOTES_VERSION = "releaseNotesVersion";
 export const SETTING_THEME_PAINTER_THEME = "themePainterTheme";