+++ src/wallmaker.js
@@ -7,11 +7,11 @@ export const makeWalls = async (state) => {
   }
   await deleteAllWalls();
   if (state.geometry) {
-    await makeWallsFromMulti(state.geometry);
+    await makeWallsFromMulti(state.config, state.geometry);
   }
-  await makeInteriorWalls(state.interiorWalls);
-  await makeDoors(state.doors);
-  await makeSecretDoors(state.secretDoors);
+  await makeInteriorWalls(state.config, state.interiorWalls);
+  await makeDoors(state.config, state.doors);
+  await makeSecretDoors(state.config, state.secretDoors);
 };
 
 const deleteAllWalls = async () => {
@@ -33,46 +33,21 @@ const deleteAllWalls = async () => {
   }
 };
 
-const makeWallsFromMulti = async (multi) => {
+const makeWallsFromMulti = async (config, multi) => {
   for (let i = 0; i < multi.getNumGeometries(); i++) {
     const poly = multi.getGeometryN(i);
-    await makeWallsFromPoly(poly);
+    await makeWallsFromPoly(config, poly);
   }
 };
 
-const wallData = (x1, y1, x2, y2) => {
-  return {
-    // From Foundry API docs:
-    // "The wall coordinates, a length-4 array of finite numbers [x0,y0,x1,y1]"
-    c: [x1, y1, x2, y2],
-    flags: {
-      "dungeon-draw": {
-        // extract string constant somewhere
-        dungeonVersion: "1.0",
-      },
-    },
-  };
-};
-
-const doorData = (x1, y1, x2, y2) => {
-  const data = wallData(x1, y1, x2, y2);
-  data.door = 1; // door
-  return data;
-};
-
-const secretDoorData = (x1, y1, x2, y2) => {
-  const data = wallData(x1, y1, x2, y2);
-  data.door = 2; // secret
-  return data;
-};
-
-const makeWallsFromPoly = async (poly) => {
+const makeWallsFromPoly = async (config, poly) => {
   const allWalls = [];
   const exterior = poly.getExteriorRing();
   const coords = exterior.getCoordinates();
   for (let i = 0; i < coords.length - 1; i++) {
     // constants.MODULE_NAME
     const data = wallData(
+      config,
       coords[i].x,
       coords[i].y,
       coords[i + 1].x,
@@ -86,6 +61,7 @@ const makeWallsFromPoly = async (poly) => {
     const coords = hole.getCoordinates();
     for (let i = 0; i < coords.length - 1; i++) {
       const data = wallData(
+        config,
         coords[i].x,
         coords[i].y,
         coords[i + 1].x,
@@ -100,10 +76,10 @@ const makeWallsFromPoly = async (poly) => {
 };
 
 /** [[x1,y1,x2,y2],...] */
-const makeInteriorWalls = async (walls) => {
+const makeInteriorWalls = async (config, walls) => {
   const allWalls = [];
   for (const wall of walls) {
-    const data = wallData(wall[0], wall[1], wall[2], wall[3]);
+    const data = wallData(config, wall[0], wall[1], wall[2], wall[3]);
     allWalls.push(data);
   }
   if (allWalls.length) {
@@ -112,10 +88,10 @@ const makeInteriorWalls = async (walls) => {
 };
 
 /** [[x1,y1,x2,y2],...] */
-const makeDoors = async (doors) => {
+const makeDoors = async (config, doors) => {
   const allDoors = [];
   for (const door of doors) {
-    const data = doorData(door[0], door[1], door[2], door[3]);
+    const data = doorData(config, door[0], door[1], door[2], door[3]);
     allDoors.push(data);
   }
   if (allDoors.length) {
@@ -124,13 +100,67 @@ const makeDoors = async (doors) => {
 };
 
 /** [[x1,y1,x2,y2],...] */
-const makeSecretDoors = async (doors) => {
+const makeSecretDoors = async (config, doors) => {
   const allDoors = [];
   for (const door of doors) {
-    const data = secretDoorData(door[0], door[1], door[2], door[3]);
+    const data = secretDoorData(config, door[0], door[1], door[2], door[3]);
     allDoors.push(data);
   }
   if (allDoors.length) {
     await canvas.scene.createEmbeddedDocuments("Wall", allDoors);
   }
+};
+
+const threeDCanvasEnabled = () => {
+  return game.settings.get(
+    constants.MODULE_NAME,
+    constants.SETTING_3DCANVAS_ENABLED
+  );
+};
+
+const wallData = (config, x1, y1, x2, y2) => {
+  const data = {
+    // From Foundry API docs:
+    // "The wall coordinates, a length-4 array of finite numbers [x0,y0,x1,y1]"
+    c: [x1, y1, x2, y2],
+    flags: {},
+  };
+  data.flags[constants.MODULE_NAME] = {};
+  data.flags[constants.MODULE_NAME][constants.FLAG_DUNGEON_VERSION] =
+    constants.DUNGEON_VERSION;
+  // Maybe set Canvas3D flags
+  if (threeDCanvasEnabled()) {
+    data.flags["levels-3d-preview"] = {
+      joinWall: true,
+      wallDepth: config.wallThickness,
+      wallTexture: config.wallTexture,
+      wallTint: config.wallTexture ? config.wallTextureTint : config.wallColor,
+    };
+  }
+  return data;
+};
+
+const doorData = (config, x1, y1, x2, y2) => {
+  const data = wallData(config, x1, y1, x2, y2);
+  data.door = 1; // door
+  // Maybe set Canvas3D flags
+  if (threeDCanvasEnabled()) {
+    data.flags["levels-3d-preview"]["joinWall"] = false;
+    // top and sides of door look like surrounding walls
+    data.flags["levels-3d-preview"]["wallSidesTexture"] = config.wallTexture;
+    data.flags["levels-3d-preview"]["wallSidesTint"] = config.wallTextureTint;
+    // actual door texture in a 2d door
+    data.flags["levels-3d-preview"]["wallTexture"] =
+      "modules/dungeon-draw/assets/textures/arena-gate-texture.webp";
+    // TODO: what do we want to do with wallTint?
+    // data.flags["levels-3d-preview"]["wallTint"] = config.doorFillColor;
+    delete data.flags["levels-3d-preview"]["wallTint"];
+  }
+  return data;
+};
+
+const secretDoorData = (config, x1, y1, x2, y2) => {
+  const data = wallData(config, x1, y1, x2, y2);
+  data.door = 2; // secret
+  return data;
 };