+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.14.0
+
+- Add support for 3D Canvas module wall flags.
+- Add "Save to Scene Background" button, useful for setting a 3D Canvas floor.
+- Change renderer to use Pixi.js polygon texture fill for floor and wall textures, rather than manually creating Sprites. This should allow small (e.g., 100x100px) textures to be performant.
+
 # 0.13.1
 
 - Fix bug where custom themes wouldn't get defaults for new config settings.